% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{count_ngram}
\alias{count_ngram}
\title{Count n-gram occurrences}
\usage{
count_ngram(tree, ngram)
}
\arguments{
\item{tree}{Suffix tree}

\item{ngram}{n-gram to search for, as a numeric or character vector.}
}
\value{
Integer n-gram count.
}
\description{
Counts n-gram occurrences in the suffix tree.
}
