% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{demo_suffix_tree}
\alias{demo_suffix_tree}
\title{Run Shiny demo}
\usage{
demo_suffix_tree(max_seq_length = 20, order_bound = 10)
}
\arguments{
\item{max_seq_length}{Maximum length of sequence that can be added each time.}

\item{order_bound}{Order bound of the suffix tree.}
}
\description{
Runs a Shiny demo for the suffixTree package.
}
