% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{get_active_order}
\alias{get_active_order}
\title{Get active order}
\usage{
get_active_order(tree)
}
\arguments{
\item{tree}{Suffix tree, as produced by \code{new_tree()}.}
}
\description{
Gets the active order of a suffix tree.
The active order is defined as the length of the
current conditioning context.
When inserting a new sequence, this will typically equal
the length of the currently inserted portion,
or the tree's order bound (whichever is less).
}
