% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-num-children.R
\name{get_num_children}
\alias{get_num_children}
\title{Get number of children}
\usage{
get_num_children(tree, order, exclude_terminals = FALSE)
}
\arguments{
\item{tree}{A suffix tree as produced by \code{new_tree}.}

\item{order}{Amount of context (e.g. 0 means no context).}
}
\value{
The number of children of the active node with order as specified,
possibly including terminal symbols.
}
\description{
Returns the number of children of a given active node,
excluding terminals if requested.
}
