% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{last_location}
\alias{last_location}
\title{Last location}
\usage{
last_location(tree)
}
\arguments{
\item{tree}{Suffix tree, as produced by \code{new_tree()}.}
}
\description{
Get the last stored location in the tree.
This is typically 0 (if no sequences have yet been stored in the tree)
or alternatively the last location in the last sequence entered into
the tree.
}
