% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{num_observed}
\alias{num_observed}
\title{Number of observed symbols}
\usage{
num_observed(tree)
}
\arguments{
\item{tree}{Suffix tree, as produced by \code{new_tree()}.}
}
\description{
Returns the number of symbols that have been entered into the tree,
including repetitions.
}
