% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{reset_active_nodes}
\alias{reset_active_nodes}
\title{Reset active nodes}
\usage{
reset_active_nodes(tree)
}
\arguments{
\item{tree}{Suffix tree, as produced by \code{new_tree()}.}
}
\description{
Resets the active nodes in a suffix tree.
This is typically used in preparation for modeling a new sequence.
}
