% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take-path.R
\name{take_path}
\alias{take_path}
\title{Take path in suffix tree}
\usage{
take_path(node, value, save = FALSE, time = NULL, pos = NULL,
  terminal = FALSE, exclude_update = FALSE)
}
\arguments{
\item{node}{Object of type \code{node}, corresponding to the current node.}

\item{value}{Name of the (possibly non-existent) target node.
Should be a scalar character.}

\item{save}{Whether to save the journey into the tree.}

\item{time}{(Numeric scalar) Ignored if \code{save} is \code{FALSE};
provides the timepoint that should be saved into the tree.}

\item{pos}{(Integerish scalar) Ignored if \code{save} is \code{FALSE};
corresponds to the 1-indexed position of the event in the training data
(across all sequences).}

\item{terminal}{(Logical scalar) Whether or not the symbol is a terminal symbol.}

\item{exclude_update}{(Logical scalar) Whether or not updates should be excluded
for this event.}
}
\value{
New node reached by taking this path (NA if no valid node found)
}
\description{
Takes a path in a suffix tree from a given node to a new node determined by
the input value.
}
