% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tst-package.R
\docType{package}
\name{tst-package}
\alias{tst}
\alias{tst-package}
\title{tst: Timestamped Suffix Trees (TSTs)}
\description{
Timestamped suffix trees (TSTs) are suffix trees that store timestamps for each state transition in the training sequence(s). Given a sequence of consecutive symbols, the TST can efficiently return  the timepoints when this sequence occurred in the training data,  the symbols that followed this sequence in the training data, and the timepoints when these continuation symbols were observed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pmcharrison/tst}
  \item Report bugs at \url{https://github.com/pmcharrison/tst/issues}
}

}
\author{
\strong{Maintainer}: Peter Harrison \email{pmc.harrison@gmail.com}

}
\keyword{internal}
