% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when.R
\name{when_context}
\alias{when_context}
\title{Find timepoints when the current context were seen in the dataset,
where the current context is defined by the tree's active nodes.}
\usage{
when_context(tree, order)
}
\arguments{
\item{tree}{A suffix tree as produced by \code{new_tree}.}

\item{order}{Amount of context (e.g. 0 means no context).}
}
\description{
Find timepoints when the current context were seen in the dataset,
where the current context is defined by the tree's active nodes.
}
