% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when.R
\name{when_continuation_terminal}
\alias{when_continuation_terminal}
\title{Find timepoints where the terminal symbol was observed after the same
context had just been entered into the tree.}
\usage{
when_continuation_terminal(tree, order)
}
\arguments{
\item{tree}{A suffix tree as produced by \code{new_tree}.}

\item{order}{Amount of context to consider (e.g. 0 means no context).}
}
\value{
\code{NULL} if no continuations found
}
\description{
Find timepoints where the terminal symbol was observed after the same
context had just been entered into the tree.
}
