% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{when_ngram}
\alias{when_ngram}
\title{Find n-gram occurrences}
\usage{
when_ngram(tree, ngram)
}
\arguments{
\item{tree}{Suffix tree}

\item{ngram}{n-gram to search for, as a numeric or character vector.}
}
\value{
Numeric vector of locations for the n-gram occurrences.
}
\description{
Finds n-gram occurrences in the suffix tree.
}
