% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{new_tree}
\alias{new_tree}
\title{Create new suffix tree}
\usage{
new_tree(order_bound = NULL, terminal = "$",
  track_update_excluded_counts = TRUE)
}
\arguments{
\item{order_bound}{Order bound for the suffix tree, i.e.
the largest n-gram that can be stored.}

\item{terminal}{Reserved terminal symbol; this character must not appear
in subsequently modeled input text.}

\item{track_update_excluded_counts}{Whether or not to track update-excluded counts.}
}
\value{
The new suffix tree.
}
\description{
Creates a new suffix tree that can subsequently be used for sequence modeling.
}
