% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when.R
\name{when_continuation}
\alias{when_continuation}
\title{Find timepoints that a given continuation was observed in the tree
from the current active node with a given order}
\usage{
when_continuation(tree, order, continuation, update_excluded = FALSE)
}
\arguments{
\item{tree}{A suffix tree as produced by \code{new_tree}.}

\item{order}{Amount of context (e.g. 0 means no context)}

\item{continuation}{Continuation of interest (should be a scalar).}

\item{update_excluded}{Boolean; whether to return update-excluded results.}
}
\value{
\code{NULL} if no continuations found
}
\description{
Find timepoints that a given continuation was observed in the tree
from the current active node with a given order
}
