% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when.R
\name{when_continuations}
\alias{when_continuations}
\title{Tabulate all continuations from the current active node with a given order}
\usage{
when_continuations(tree, order, update_excluded = FALSE)
}
\arguments{
\item{tree}{A suffix tree as produced by \code{new_tree}.}

\item{order}{Amount of context to consider (e.g. 0 means no context).}

\item{update_excluded}{Boolean; whether to return update-excluded results.}
}
\description{
Tabulate all continuations from the current active node with a given order
}
